backend/requirements_new.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
alabaster==1.0.0
alembic==1.16.4
annotated-types==0.7.0
anyio==3.7.1
apscheduler==3.11.0
arrow==1.3.0
asgiref==3.8.1
asyncpg>=0.30.0
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bibtexparser==1.4.3
boto3==1.40.47
cachetools==5.5.2
certifi==2025.4.26
cffi>=1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.2
dataclasses-json==0.6.7
dateparser==1.2.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
docutils==0.21.2
dotenv==0.9.9
duckduckgo_search==8.0.1
ecdsa==0.19.1
email-service==2.0.0
et_xmlfile==2.0.0
fake-useragent==2.2.0
fastapi==0.115.12
filelock==3.18.0
free_proxy==1.1.3
frozenlist==1.6.0
fsspec==2025.3.2
geographiclib==2.0
geopy==2.4.1
google-ai-generativelanguage==0.6.10
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
greenlet==3.2.1
groq==0.11.0
grpcio>=1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.1
huggingface-hub==0.30.2
idna==3.10
imagesize==1.4.1
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
lxml>=5.4.0
Mako==1.3.10
markdown2==2.5.3
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
networkx==3.4.2
openpyxl==3.1.5
orjson==3.11.2
outcome==1.3.0.post0
packaging==25.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
primp==0.15.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.1.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pyee==13.0.0
Pygments==2.19.2
pymongo==4.13.2
pyparsing==3.2.3
pypdfium2==4.30.1
PySocks==1.7.1
python-dotenv==1.1.0
python-http-client==3.3.7
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests_cache==1.2.1
requests-file==2.1.0
requests-toolbelt==1.0.0
roman-numerals-py==3.1.0
rsa==4.9.1
safetensors==0.5.3
scholarly==1.7.11
sendgrid==6.4.8
sentence-transformers==4.1.0
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy>=3.8.0
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.35
starkbank-ecdsa==2.2.0
starlette==0.46.2
stem==1.8.2
sympy==1.13.1
tenacity==9.1.2
tldextract==5.3.0
tokenizers==0.21.1
tqdm==4.67.1
transformers>=4.40.0
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zstandard==0.23.0
